updates, deletes and adds commands
