+++ electron/commandsUtil/commandsUtil.js
@@ -6,31 +6,9 @@ var loadPhrases = require('../utils/loaders').loadPhrases;
 var prefixTrie = require('../match/prefixTrie');
 var save = require('../utils/utils').save;
 var write = require('../utils/utils').write;
+var get = require('../utils/utils').get;
+var lowerCaseProps = require('../utils/utils').lowerCaseProps;
 var parseCommands = require('../match/parseCommands').parseCommands;
-
-
-var get = function (name) {
-  return JSON.parse(localStorage.getItem(name));
-};
-
-var lowerCaseProps = function (obj) {
-  var newObj = {};
-  for (var key in obj) {
-    newObj[key.toLowerCase()] = obj[key];
-  }
-  return newObj;
-};
-
-// var readAsync = function (filePath, cb) {
-//   fs.readFile(filePath, 'utf8', function (err, data) {
-//     if (err) {
-//       cb(err);
-//     } else {
-//       cb(null, JSON.parse(data));
-//     }
-//   })
-// };
-
 
 module.exports.saveCommands = function (obj) {
   console.log('SAVING', obj.rawCommands);
@@ -38,6 +16,10 @@ module.exports.saveCommands = function (obj) {
     obj = JSON.stringify(obj);
   }
   save('Commands', obj);
+};
+
+module.exports.getCommands = function () {
+  return get('Commands');
 };
 
 module.exports.loadPackage = function (commandsPath) {
@@ -52,44 +34,7 @@ module.exports.loadPackage = function (commandsPath) {
   module.exports.saveCommands(commandObj);
 };
 
-// module.exports.loadPackageAsync = function (commandsPath, cb) {
-//   var commandObj = {};
-//   readAsync(commandsPath, function (err, data) {
-//     if (err) {
-//       cb(err);
-//     } else {
-//       var rawCommands = data;
-//       commandObj.rawCOmmands = rawCommands;
-//       commandObj.parsedCommands = parseCommands(rawCommands);
-//       commandObj.commandsPath = commandsPath;
-//       commandObj.phrasesPath = commandsPath.replace('commands.', 'phrases.');
-//       commandObj.phrases = loadPhrases(commandObj.phrasesPath, commandObj.rawCommands);
-//       prefixTrie.build(Object.keys(commandObj.parsedCommands.argCommands));
-//       module.exports.saveCommands(commandObj);
-//       cb(null, commandObj);
-//     }
-//   })
-// };
-
-
-module.exports.getCommands = function () {
-  return get('Commands');
-};
-
-
-// module.exports.addCommand = function (command) {
-//   console.log('I GOT A COMMAND FROM THE VIEWS');
-//   var newCommandsObj = _.extend({}, module.exports.getCommands());
-//   newCommandsObj.rawCommands = lowerCaseProps(_.extend(module.exports.getCommands().rawCommands, command));
-//   newCommandsObj.parsedCommands = parseCommands(newCommandsObj.rawCommands);
-//   module.exports.saveCommands(newCommandsObj);
-//   write(newCommandsObj.commandsPath, newCommandsObj.rawCommands);
-//   module.exports.addPhrase(Object.keys(command)[0], Object.keys(command)[0]);
-// };
-//
-
 module.exports.updateCommands = function (command) {
-  console.log('I GOT A COMMAND FROM THE VIEWS');
   var newCommandsObj = _.extend({}, module.exports.getCommands());
   newCommandsObj.rawCommands = lowerCaseProps(command);
   newCommandsObj.parsedCommands = parseCommands(newCommandsObj.rawCommands);
@@ -106,21 +51,6 @@ module.exports.delCommand = function (command) {
   module.exports.saveCommands(commandsObj);
   write(commandsObj.commandsPath, commandsObj.rawCommands);
   write(commandsObj.phrasesPath, commandsObj.phrases);
-};
-
-
-module.exports.updateCommand = function (command, action, oldCommand) {
-  var commandsObj = module.exports.getCommands();
-  if (oldCommand === command) {
-    commandsObj.rawCommands[command] = action;
-  } else {
-    delete commandsObj.rawCommands[oldCommand];
-    delete commandsObj.phrases[oldCommand];
-    module.exports.addPhrase(command, command);
-    commandsObj.rawCommands[command] = action;
-  }
-  module.exports.saveCommands(commandsObj);
-  write(newCommandsObj.commandsPath, newCommandsObj.rawCommands);
 };
 
 module.exports.addPhrase = function (correctCommand, userCommand) {
