+++ electron/packages/newphrases.json
@@ -1,1 +1,1 @@
-{"say hello":["say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello","say hello"],"enhance":["enhance"],"zoom out":["zoom out"],"dance for me":["dance for me"],"new command":["new command"],"old command":["old command"],"jarvis":["jarvis"],"dehance":["dehance"],"do awesome things":["do awesome things"],"free willy":["free willy"],"what time is it":["what time is it"],"":[""],"check the":["check the"],"kyle cho pro tip":["kyle cho pro tip","kyle joe pro tip"],"dim screen":["dim screen"],"open":["open"],"volume down":["volume down"],"volume up":["volume up"],"volume off":["volume off"],"say goodbye":["say goodbye","say goodbye"],"say":["say","say","say","say","say","say","say","say","say","say","say","say","say","say","say"]}
+{"volume up":["volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up","volume up"],"say":["say"],"enhance":["enhance"],"zoom out":["zoom out"],"dance for me":["dance for me"],"jarvis":["jarvis"],"dehance":["dehance"],"do awesome things":["do awesome things"],"free willy":["free willy"],"what time is it":["what time is it"],"check the":["check the"],"kyle cho pro tip":["kyle cho pro tip","kyoto protocol"],"dim screen":["dim screen"],"open":["open"],"volume down":["volume down"],"volume off":["volume off"]}
